+++ lib/src/base_state.dart
@@ -5,7 +5,6 @@ import 'package:apn_state/src/event_bus_event.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 
-
 enum ViewState { Idle, Busy }
 
 abstract class BaseState<V> extends ChangeNotifier {
@@ -23,15 +22,23 @@ abstract class BaseState<V> extends ChangeNotifier {
 
   V convertError(dynamic e);
 
+  /// Emit an event that can be picked up by all other states
+  @protected
   void emit(EventBusEvent event) => EventBus.emit(event);
+
+  /// Listen to a specific type of event. This automatically closes
+  /// the listener when the state is disposed
+  @protected
   StreamSubscription<T> listen<T extends EventBusEvent>(void onListen(T)) {
-     final subscription = EventBus.on<T>(onListen);
-     _subscriptions.add(subscription);
-     return subscription;
+    final subscription = EventBus.on<T>(onListen);
+    _subscriptions.add(subscription);
+    return subscription;
   }
 
+  /// Make sure to implement the super.dispose() when
+  /// overriding this method
   @override
-  void dispose(){
+  void dispose() {
     _subscriptions.forEach((sub) => sub.cancel());
     super.dispose();
   }
@@ -47,23 +54,26 @@ abstract class BaseState<V> extends ChangeNotifier {
     notifyListeners();
   }
 
-  /// To shut up the invalid warning of "notifyListeners can only be called within the ChangeNotifier class"
+  /// We allow calling the notifyListeners on our state
+  /// objects. So we publicly expose it (in stead of relying
+  /// on the parents @protected)
   void notifyListeners() => super.notifyListeners();
 
-  Future<T> process<T>(ValueGetter<Future<T>> callback) async {
+  /// Asynchronously handle the callback and automatically set the state loading.
+  Future<T> process<T>(ValueGetter<Future<T>> callback, [handleLoading = true]) async {
     // * Clear previous error
     error = null;
     T response;
 
     try {
-      setState(ViewState.Busy);
+      if (handleLoading) setState(ViewState.Busy);
       response = await callback();
     } catch (e) {
       print(e);
       error = convertError(e);
     }
 
-    setState(ViewState.Idle);
+    if (handleLoading) setState(ViewState.Idle);
     return response;
   }
 }